Renamed username tab state files
