+++ lib/state_management/usernames/usernames_tab_notifier.dart
@@ -6,7 +6,7 @@ import 'package:anonaddy/models/username/username_model.dart';
 import 'package:anonaddy/services/data_storage/offline_data_storage.dart';
 import 'package:anonaddy/services/username/username_service.dart';
 import 'package:anonaddy/state_management/lifecycle/lifecycle_state_manager.dart';
-import 'package:anonaddy/state_management/usernames/usernames_state.dart';
+import 'package:anonaddy/state_management/usernames/usernames_tab_state.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 final usernameStateNotifier =
